Split server from main app entry point.
